refactor: relocate Conversions out of internal.schema to common.schema.types
Conversions only depends on Type / Types — both already living under
common.schema.types — so it relocates verbatim alongside them. The single
external caller (SparkHoodieTableFileIndex) updates its import.
